Extract `start_global_dispute` tests
